requirements.txt
bcrypt>=4.0.0
pandas==2.1.4
numpy>=1.26.0
python-dotenv==1.0.0
scikit-learn>=1.3.2
joblib>=1.3.2
Flask==3.1.2
flask-cors==6.0.1
gunicorn>=23.0.0
pypdf==5.4.0
Pillow==11.2.1
pytesseract==0.3.13
pydantic>=2.0.0
structlog>=24.0.0
PyMySQL>=1.1.1
pytest>=8.0.0
pytest-flask>=1.3.0
google-genai>=1.0.0
requests>=2.31.0
